feat(storybook): update configuration and fix TypeScript setup
